# docs/API/openapi/api-v1.openapi.yaml
openapi: 3.1.0
info:
  title: Spark API v1
  version: "1.0.0"
  description: >
    Spark's client-neutral, capability-scoped REST API. Hand-written from
    routes/api.php and the underlying controllers on branch
    feature/api-mobile-mcp-parity — not generated, so keep it updated
    alongside docs/API/API_v1.md when routes change. See
    docs/API/README.md for the cross-surface capability model.
...
servers:
  - url: /api/v1
security:
  - sparkToken: []
...
paths:
...
  /tags:
    get:
      operationId: listTags
      tags: [data]
      x-spark-ability: data:read
      summary: Cursor-paginated list of the user's tags
      parameters:
        - { name: q, in: query, schema: { type: string, maxLength: 255 } }
        - { name: cursor, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, maximum: 100, default: 30 } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Tag' } }
                  next_cursor: { type: [string, "null"] }
                  has_more: { type: boolean }
...
  /tags/{id}:
    get:
      operationId: getTag
      tags: [data]
      x-spark-ability: data:read
      summary: A single tag plus the items tagged with it
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
        - { name: cursor, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, maximum: 100, default: 30 } }
      responses:
        "200": { description: OK }
        "404": { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    sparkToken:
      type: http
      scheme: bearer
      description: >
        Laravel Sanctum personal access token. Each operation's x-spark-ability
        extension names the ability the token must carry
        (App\Support\SparkAbility::allows) — OpenAPI has no native primitive
        for Sanctum's per-operation string abilities.
...
  responses:
    NotFound:
      description: Not found or not owned by the caller
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      properties:
        message: { type: string }
        hint: { type: string }
        errors: { type: object }
...
    Tag:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        type: { type: [string, "null"] }
        events_count: { type: integer }
        objects_count: { type: integer }
        total_count: { type: integer }
...